PYIC-8399: refactor shared ApiGatewayProxyEvent functions to helper class to be used by other lambdas in libs package
