+++ di-ipv-cimit-stub/lib/build.gradle
@@ -13,7 +13,8 @@ dependencies {
 			"software.amazon.lambda:powertools-parameters:1.20.2",
 			"software.amazon.awssdk:kms:2.31.54",
 			"org.apache.logging.log4j:log4j-api:2.24.3",
-			"org.apache.logging.log4j:log4j-core:2.24.3"
+			"org.apache.logging.log4j:log4j-core:2.24.3",
+			"com.amazonaws:aws-lambda-java-events:3.15.0"
 
 	testImplementation "org.junit.jupiter:junit-jupiter:5.13.1",
 			"org.mockito:mockito-junit-jupiter:5.18.0",
