+++ di-ipv-cimit-stub/lib/src/main/java/uk/gov/di/ipv/core/library/exceptions/FailedToParseRequestException.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.getcontraindicatorcredential.exceptions;
+package uk.gov.di.ipv.core.library.exceptions;
 
 public class FailedToParseRequestException extends Exception {
     public FailedToParseRequestException(String message) {super(message);}
